+++ ExampleApp/RoomListScreen.js
@@ -1,6 +1,7 @@
 import React from 'react';
-import { RoomList } from '..';
+import { RoomList, matrix } from '..';
 import { screens } from './App';
+import { View, TouchableOpacity } from 'react-native';
 
 export default function RoomListScreen({ goToScreen, setCurrentRoom }) {
   const handleOnRowPress = room => {
@@ -8,5 +9,27 @@ export default function RoomListScreen({ goToScreen, setCurrentRoom }) {
     goToScreen(screens.CHAT);
   };
 
-  return <RoomList onRowPress={handleOnRowPress} />;
+  const createChat = async () => {
+    const room = await matrix.createRoom();
+    console.log(room);
+  };
+
+  return (
+    <>
+      <View
+        style={{
+          width: '100%',
+          height: 50,
+          backgroundColor: 'pink',
+          justifyContent: 'center',
+          alignItems: 'center',
+        }}>
+        <TouchableOpacity
+          onPress={createChat}
+          style={{ backgroundColor: 'blue', width: 200, height: 25 }}
+        />
+      </View>
+      <RoomList onRowPress={handleOnRowPress} />
+    </>
+  );
 }